requirements.txt
pyautogui
pillow
opencv-python
pyscreeze
pydirectinput
requests
keyboard
pywin32
